+++ frontend/catalyst/pennylane_extensions.py
@@ -394,10 +394,11 @@ def _ensure_differentiable(f: DifferentiableLike) -> Differentiable:
 
 
 def _make_jaxpr_check_differentiable(f: Differentiable, grad_params: GradParams, *args) -> Jaxpr:
-    """Gets the jaxpr of a differentiable function. Perform the required additional checks."""
+    """Gets the jaxpr of a differentiable function. Perform the required additional checks and
+    return the output tree."""
     method = grad_params.method
-    jaxpr = jax.make_jaxpr(f)(*args)
-
+    jaxpr, shape = jax.make_jaxpr(f, return_shape=True)(*args)
+    _, out_tree = tree_flatten(shape)
     assert len(jaxpr.eqns) == 1, "Expected jaxpr consisting of a single function call."
     assert jaxpr.eqns[0].primitive == func_p, "Expected jaxpr consisting of a single function call."
 
@@ -423,7 +424,7 @@ def _make_jaxpr_check_differentiable(f: Differentiable, grad_params: GradParams,
 
     _verify_differentiable_child_qnodes(jaxpr, method)
 
-    return jaxpr
+    return jaxpr, out_tree
 
 
 def _verify_differentiable_child_qnodes(jaxpr, method):
@@ -522,6 +523,23 @@ def _check_grad_params(
     return GradParams(method, scalar_out, h, argnum)
 
 
+def _unflatten_derivatives(results, out_tree, argnum, num_results):
+    """Unflatten the flat list of derivatives results given the out tree."""
+    num_trainable_params = len(argnum) if isinstance(argnum, list) else 1
+    results_final = []
+    for i in range(0, num_results):
+        intermediate_results = results[
+            i * num_trainable_params : i * num_trainable_params + num_trainable_params
+        ]
+        if not (isinstance(argnum, int) or argnum is None):
+            intermediate_results = tuple(intermediate_results)
+        else:
+            intermediate_results = intermediate_results[0]
+        results_final.append(intermediate_results)
+    results_final = tree_unflatten(out_tree, results_final)
+    return results_final
+
+
 class Grad:
     """An object that specifies that a function will be differentiated.
 
@@ -547,16 +565,19 @@ class Grad:
         Args:
             args: the arguments to the differentiated function
         """
+
         if EvaluationContext.is_tracing():
             fn = _ensure_differentiable(self.fn)
             grad_params = _check_grad_params(*self.grad_params)
-
-            jaxpr = _make_jaxpr_check_differentiable(fn, grad_params, *args)
+            jaxpr, out_tree = _make_jaxpr_check_differentiable(fn, grad_params, *args)
 
             args_data, _ = tree_flatten(args)
 
             # It always returns list as required by catalyst control-flows
             results = grad_p.bind(*args_data, jaxpr=jaxpr, fn=fn, grad_params=grad_params)
+            results = _unflatten_derivatives(
+                results, out_tree, self.grad_params.argnum, len(jaxpr.out_avals)
+            )
         else:
             if argnums := self.grad_params.argnum is None:
                 argnums = 0
@@ -573,7 +594,7 @@ def grad(f: DifferentiableLike, *, method=None, h=None, argnum=None):
 
     This function allows the gradient of a hybrid quantum-classical function to be computed within
     the compiled program. Outside of a compiled function, this function will simply dispatch to its
-    JAX counterpart ``jax.grad``.
+    JAX counterpart ``jax.grad``. The function ``f`` can return any pytree-like shape.
 
     .. warning::
 
@@ -703,7 +724,7 @@ def jacobian(f: DifferentiableLike, *, method=None, h=None, argnum=None):
 
     This function allows the Jacobian of a hybrid quantum-classical function to be computed within
     the compiled program. Outside of a compiled function, this function will simply dispatch to its
-    JAX counterpart ``jax.jacobian``.
+    JAX counterpart ``jax.jacobian``. The function ``f`` can return any pytree-like shape.
 
     Args:
         f (Callable): a function or a function object to differentiate
@@ -756,8 +777,8 @@ def jacobian(f: DifferentiableLike, *, method=None, h=None, argnum=None):
             return g(x)
 
     >>> workflow(jnp.array([2.0, 1.0]))
-    array([[-1.32116540e-07,  1.33781874e-07],
-           [-4.20735506e-01,  4.20735506e-01]])
+    array([[ 3.48786850e-16 -4.20735492e-01]
+           [-8.71967125e-17  4.20735492e-01]])
     """
     scalar_out = False
     return Grad(f, GradParams(method, scalar_out, h, argnum))
@@ -769,7 +790,8 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
 
     This function allows the Jacobian-vector Product of a hybrid quantum-classical function to be
     computed within the compiled program. Outside of a compiled function, this function will simply
-    dispatch to its JAX counterpart ``jax.jvp``.
+    dispatch to its JAX counterpart ``jax.jvp``. The function ``f`` can return any pytree-like
+    shape.
 
     Args:
         f (Callable): Function-like object to calculate JVP for
@@ -782,8 +804,8 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
         h (float): the step-size value for the finite-difference (``"fd"``) method
         argnum (Union[int, List[int]]): the params' indices to differentiate.
 
-    Returns (Tuple[Array]):
-        Return values of ``f`` paired with the JVP values.
+    Returns:
+        Tuple[Any]: Return values of ``f`` paired with the JVP values.
 
     Raises:
         TypeError: invalid parameter types
@@ -804,8 +826,7 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
     >>> x = jnp.array([0.1, 0.2])
     >>> tangent = jnp.array([0.3, 0.6])
     >>> jvp(x, tangent)
-    [array([0.09983342, 0.04      , 0.02      ]),
-    array([0.29850125, 0.24000006, 0.12      ])]
+    (array([0.09983342, 0.04      , 0.02      ]), array([0.29850125, 0.24      , 0.12      ]))
 
     **Example 2 (argnum usage)**
 
@@ -829,7 +850,7 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
     >>> params = jnp.array([[0.54, 0.3154], [0.654, 0.123]])
     >>> dy = jnp.array([[1.0, 1.0], [1.0, 1.0]])
     >>> workflow(params, dy)
-    [array(0.78766064), array(-0.7011436)]
+    (array(0.78766064), array(-0.7011436))
     """
 
     def check_is_iterable(x, hint):
@@ -844,9 +865,18 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
         fn = _ensure_differentiable(f)
         grad_params = _check_grad_params(method, scalar_out, h, argnum)
 
-        jaxpr = _make_jaxpr_check_differentiable(fn, grad_params, *params)
+        jaxpr, out_tree = _make_jaxpr_check_differentiable(fn, grad_params, *params)
 
         results = jvp_p.bind(*params, *tangents, jaxpr=jaxpr, fn=fn, grad_params=grad_params)
+
+        midpoint = len(results) // 2
+        func_res = results[:midpoint]
+        jvps = results[midpoint:]
+
+        func_res = tree_unflatten(out_tree, func_res)
+        jvps = tree_unflatten(out_tree, jvps)
+        results = (func_res, jvps)
+
     else:
         results = jax.jvp(f, params, tangents)
 
@@ -859,7 +889,8 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
 
     This function allows the Vector-Jacobian Product of a hybrid quantum-classical function to be
     computed within the compiled program. Outside of a compiled function, this function will simply
-    dispatch to its JAX counterpart ``jax.vjp``.
+    dispatch to its JAX counterpart ``jax.vjp``. The function ``f`` can return any pytree-like
+    shape.
 
     Args:
         f(Callable): Function-like object to calculate JVP for
@@ -872,8 +903,8 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
         h (float): the step-size value for the finite-difference (``"fd"``) method
         argnum (Union[int, List[int]]): the params' indices to differentiate.
 
-    Returns (Tuple[Array]):
-        Return values of ``f`` paired with the JVP values.
+    Returns:
+        Tuple[Any]): Return values of ``f`` paired with the VJP values.
 
     Raises:
         TypeError: invalid parameter types
@@ -894,8 +925,7 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
     >>> x = jnp.array([0.1, 0.2])
     >>> dy = jnp.array([-0.5, 0.1, 0.3])
     >>> vjp(x, dy)
-    [array([0.09983342, 0.04      , 0.02      ]),
-    array([-0.43750208,  0.07000001])]
+    (array([0.09983342, 0.04      , 0.02      ]), (array([-0.43750208,  0.07      ]),))
     """
 
     def check_is_iterable(x, hint):
@@ -910,9 +940,21 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
         fn = _ensure_differentiable(f)
         grad_params = _check_grad_params(method, scalar_out, h, argnum)
 
-        jaxpr = _make_jaxpr_check_differentiable(fn, grad_params, *params)
+        jaxpr, out_tree = _make_jaxpr_check_differentiable(fn, grad_params, *params)
+
+        cotangents, _ = tree_flatten(cotangents)
 
         results = vjp_p.bind(*params, *cotangents, jaxpr=jaxpr, fn=fn, grad_params=grad_params)
+
+        func_res = results[: len(jaxpr.out_avals)]
+        vjps = results[len(jaxpr.out_avals) :]
+        func_res = tree_unflatten(out_tree, func_res)
+        # We do not change the shape of the VJP as it is the same as the parameters not
+        # as the output
+        # TODO: update when we add support for pytrees params
+
+        results = tuple([func_res, tuple(vjps)])
+
     else:
         primal_outputs, vjp_fn = jax.vjp(f, *params)